+++ app/Http/Controllers/Auth/RegisterController.php
@@ -6,6 +6,7 @@ use App\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Foundation\Auth\RegistersUsers;
 
 class RegisterController extends Controller
@@ -26,12 +27,18 @@ class RegisterController extends Controller
      * The user has been registered.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  mixed  $user
-     * @return mixed
+     * @param  \App\User  $user
+     * @return \Illuminate\Http\JsonResponse
      */
-    protected function registered(Request $request, $user)
+    protected function registered(Request $request, User $user)
     {
-        return $user;
+        if ($user instanceof MustVerifyEmail) {
+            $user->sendEmailVerificationNotification();
+
+            return response()->json(['status' => trans('verification.sent')]);
+        }
+
+        return response()->json($user);
     }
 
     /**
@@ -53,7 +60,7 @@ class RegisterController extends Controller
      * Create a new user instance after a valid registration.
      *
      * @param  array  $data
-     * @return User
+     * @return \App\User
      */
     protected function create(array $data)
     {